+++ CdIts.Caffoa.Abstractions/ICaffoaConverter.cs
@@ -1,3 +1,5 @@
+using Microsoft.Extensions.Primitives;
+
 namespace Caffoa;
 
 /// <summary>
@@ -28,7 +30,7 @@ public interface ICaffoaConverter
     /// <param name="parameter">the input string</param>
     /// <param name="parameterName">the name of the variable as defined in the openapi spec</param>
     DateTime ParseDateTime(string parameter, string parameterName);
-    
+
     /// <summary>
     /// Should parse openapi uuid format string into a Guid
     /// </summary>
@@ -37,11 +39,17 @@ public interface ICaffoaConverter
     Guid ParseGuid(string parameter, string parameterName);
     
     /// <summary>
+    /// Should parse enum types from string to the required enum type
+    /// </summary>
+    /// <param name="parameter">the input string</param>
+    /// <param name="parameterName">the name of the variable as defined in the openapi spec</param>
+    /// <param name="ignoreCase">set to false to make case-sensitive string-to-enum conversions</param>
+    T ParseEnum<T>(string parameter, string parameterName, bool ignoreCase = true)  where T : Enum;
+    
+    /// <summary>
     /// Should parse all other items supported by openapi (e.g. int, uint, long, ulong, Guid)
     /// </summary>
     /// <param name="parameter">the input string</param>
     /// <param name="parameterName">the name of the variable as defined in the openapi spec</param>
     T Parse<T>(string parameter, string parameterName);
-
-    
 }