+++ CdIts.Caffoa.Cli/Extensions.cs
@@ -141,6 +141,12 @@ public static class Extensions
     {
         return property.Enums.Count > 0 && !property.TypeName.StartsWith("double") && !property.TypeName.StartsWith("float");
     }
+    
+    public static bool CanBeEnum(this OpenApiSchema schema)
+    {
+        return schema.Enum.Count > 0 && !schema.Type.StartsWith("double") && !schema.Type.StartsWith("float");
+    }
+    
     public static bool CanBeStringEnum(this PropertyData property)
     {
         return property.Enums.Count > 0;