+++ CdIts.Caffoa.Cli/Model/ParameterObject.cs
@@ -11,6 +11,7 @@ public class ParameterObject
 
     public string? DefaultValue { get; set; }
     public bool Required { get; set; }
+    public bool IsEnum { get; set; }
 
     public ParameterObject(string name, string typeName, string description, bool inQuery)
     {