+++ CdIts.Caffoa.Cli/Model/SchemaItem.cs
@@ -2,12 +2,19 @@ namespace CdIts.Caffoa.Cli.Model;
 
 public class SchemaItem
 {
+    public enum ObjectType
+    {
+        Regular,
+        StringEnum,
+        IntEnum
+    };
     public SchemaItem(string name, string className)
     {
         Name = name;
         ClassName = className;
     }
 
+    public  ObjectType Type { get; set; } = ObjectType.Regular;
     public string Name { get; }
     public string ClassName { get; }
     public InterfaceModel? Interface { get; set; }
@@ -16,4 +23,5 @@ public class SchemaItem
     public string? Description { get; set; }
     public bool AdditionalPropertiesAllowed { get; set; }
     public List<string> SubItems { get; set; } = new();
+    public List<string?> Enums { get; set; } = new();
 }